Fix username migration under certain circumstances
Heroku's deploy was failing due to username not having a default value
as well as username having a null constraint. Upon adding this column,
the database didn't know what value the username field should inherit so
the migration failed. This wasn't an issue in dev envs because on this
migration run, I assume that there were no users at first.